+++ include/zmsg/zmsg_fusion_splice.hpp
@@ -38,7 +38,7 @@ public:
 
 	BOOL ConeFS;
 	ULONG ConeFSWaitTime;
-	ULONG ConeFSSpeed;
+	DOUBLE ConeFSSpeed;
 	ULONG ConeFSStretchLength;
 
 	loss_estimate_mode_t LossEstimationMode;
